'strlen' returns 'size_t'; perform arithmetic at 'size_t'.
SVN r7600

+++ runtime/gc/done.c
@@ -1,4 +1,5 @@
-/* Copyright (C) 1999-2008 Henry Cejtin, Matthew Fluet, Suresh
+/* Copyright (C) 2012 Matthew Fluet.
+ * Copyright (C) 1999-2008 Henry Cejtin, Matthew Fluet, Suresh
  *    Jagannathan, and Stephen Weeks.
  * Copyright (C) 1997-2000 NEC Research Institute.
  *
@@ -6,10 +7,10 @@
  * See the file MLton-LICENSE for details.
  */
 
-static void displayCol (FILE *out, int width, const char *s) {
-  int extra;
-  int i;
-  int len;
+static void displayCol (FILE *out, size_t width, const char *s) {
+  size_t extra;
+  size_t i;
+  size_t len;
 
   len = strlen (s);
   if (len < width) {
